Initial scripts for scraping data and structuring
The SQLite database is the canonical record.  I'm also keeping the data
in a CSV file, which will be better for uploading to the git repository,
since the deltas will be small (updates to a binary file won't be
optimized).  I'm also keeping the current info in JSONs at sensible
endpoints, so I can just request those if I need the current status.